+++ _custom_build/build.py
@@ -1,0 +1,7 @@
+from __future__ import annotations
+
+from distutils.command.build import build as orig_build
+
+
+class build(orig_build):
+    sub_commands = orig_build.sub_commands + [('fetch_binaries', None)]
